Refactor strings.xml and EditActivity
I've moved every string into strings.xml, in preparation to the italian
release. I've not done internationalization yet.

I've also tried to improve the tasks' editing, but it doesn't work yet.
Actually it's broken. My idea is to add an int "id" to the ToDoRow entity,
so I can put that id into a bundle and pass it from the MainActivity to the EditActivity.

Signed-off-by: simonewebdesign <hello@simonewebdesign.it>